pyproject.toml
[project]
name = "graph-rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.4",
    "flask>=3.1.3",
    "kuzu>=0.11.3",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
